feat: themable Presenter
