+++ packages/components/src/MDXDeck.js
@@ -4,6 +4,7 @@ import { Router, globalHistory, navigate } from '@reach/router'
 import { Global } from '@emotion/core'
 import { Swipeable } from 'react-swipeable'
 import merge from 'lodash.merge'
+import defaultTheme from '@mdx-deck/themes/base'
 
 import Provider from './Provider'
 import Slide from './Slide'
@@ -41,23 +42,16 @@ const getIndex = ({ basepath }) => {
   return index
 }
 
+const mergeThemes = themes =>
+  themes.reduce(
+    (acc, theme) =>
+      typeof theme === 'function' ? theme(acc) : merge(acc, theme),
+    {}
+  )
+
 const mergeState = (state, next) =>
   merge({}, state, typeof next === 'function' ? next(state) : next)
 const useState = init => useReducer(mergeState, init)
-
-const getWrapper = mode => {
-  switch (mode) {
-    case PRESENTER:
-      return Presenter
-    case OVERVIEW:
-      return Overview
-    case GRID:
-      return Grid
-    default:
-      return BaseWrapper
-      break
-  }
-}
 
 export const MDXDeckContext = React.createContext()
 
@@ -83,13 +77,28 @@ export const MDXDeckState = ({ children }) => {
 }
 
 export const MDXDeck = props => {
-  const { slides, basepath } = props
+  const { slides, basepath, theme: baseTheme, themes = [] } = props
   const { state, setState } = useDeckState(MDXDeckContext)
+  const theme = mergeThemes([defaultTheme, baseTheme, ...themes])
 
   const index = getIndex(props)
 
   const getMeta = i => {
     return state.metadata[i] || {}
+  }
+
+  const getWrapper = mode => {
+    switch (mode) {
+      case PRESENTER:
+        return theme.Presenter || Presenter
+      case OVERVIEW:
+        return Overview
+      case GRID:
+        return Grid
+      default:
+        return BaseWrapper
+        break
+    }
   }
 
   const register = (index, meta) => {
@@ -155,7 +164,7 @@ export const MDXDeck = props => {
   const Wrapper = getWrapper(state.mode)
 
   return (
-    <Provider {...props} {...context}>
+    <Provider {...props} {...context} theme={theme}>
       <Catch>
         <Style {...context} />
         <Keyboard {...props} {...context} />
