+++ packages/components/src/Provider.js
@@ -1,23 +1,13 @@
 import React from 'react'
 import { ThemeProvider } from 'emotion-theming'
-import merge from 'lodash.merge'
 import { HeadProvider, UserHead } from './Head'
 import { MDXProvider } from '@mdx-js/react'
-import defaultTheme from '@mdx-deck/themes/base'
 import mdxComponents from './mdx-components'
 
 const DefaultProvider = props => <>{props.children}</>
 
-const mergeThemes = themes =>
-  themes.reduce(
-    (acc, theme) =>
-      typeof theme === 'function' ? theme(acc) : merge(acc, theme),
-    {}
-  )
-
 export const Provider = props => {
-  const { headTags, theme: baseTheme, themes = [], mdx } = props
-  const theme = mergeThemes([defaultTheme, baseTheme, ...themes])
+  const { headTags, theme, mdx } = props
   const {
     Provider: UserProvider = DefaultProvider,
     components: themeComponents = {},
